Implement storing events
If a 'started' event comes in, store the 'type'
i.e. the topic and the timestamp.

Handle this using a go channel to prevent
synchronization issues.

This is pretty basic and relies on the data coming
in to be correct at this point. Adding proper
validation is something that could be improved on.