Add event on started at message
Add a debug config using delve. This can be
accessed via air -c .air.debug.toml

Add a launch.json for vscode debugging

Update the nulltime type to use the sql.nulltime
base type instead of time.time. This should work
better with the db timestamps, and allow us to set
a null timestamp in the db instead of the zero value
which is more _technically_ correct

Get event updates working

Update the finish existing event function to
update an existing event with the finished at time
when it receives a finish event.

Add the update function to the event service

Also tweak the nulltime scan function to return
an invalid time if the time is null in the db