Add ntfy service
Add a new manager and service to handle sending
ntfy messages to users.

Wire it in to the main subscription logic, so
now when a finished event is received, the ntfy
service will send a message to the user.

Add dockerfile

This uses multistage builds to create a small image with the compiled binary.