Fixed views not working
Fixed an issue with views not working in the container because they weren't included in the build.
Embed the views and static content using the go embed system so they are included in the binary directly.

This should also make a single binary deployment possible without needing docker

Move the template and static files into the internal package for simplicity sake