Potentially fix reconnection failures
According to old mate claude, the problem around the reconnects failing to receive events is because the library handles the reconnection automatically, but doesn't handle setting up the subscriptions to the topic after a reconnect.
So i've updated it subscribe to the topic via a connection handler callback instead of the one call when the app is initialized.

Hopefully this will mean that any time the connection is (re)established, the subscription will be too.

Also update the air file, since tailwind is up to v4 now and changes how it works, so without doing some upgrades we need to specify the version in the command